implemented showMethods to display the menues in Main class

+++ src/net/htlgkr/GumpelmeierA21044/pos3/hue2/Number.java
@@ -4,7 +4,7 @@ public class Number{
     private double a;
     private double b;
 
-    public Number(double a, int b) {
+    public Number(double a, double b) {
         this.a = a;
         this.b = b;
     }
